5-tests/requirements.txt
aiohttp==3.7.3
aiohttp-apispec==2.2.1
alembic==1.5.4
apispec==3.3.2
appdirs==1.4.4
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
black==20.8b1
chardet==3.0.4
click==7.1.2
coverage==5.4
gino==1.0.1
idna==3.1
iniconfig==1.1.1
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
multidict==5.1.0
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-aiohttp==0.3.0
pytest-cov==2.11.1
pytest-mock==3.5.1
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.4.1
regex==2020.11.13
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
webargs==5.5.3
yarl==1.6.3
